BAVL-516 reduce cache times. (#306)
