+++ helm_deploy/values-dev.yaml
@@ -16,7 +16,6 @@ generic-service:
     API_BASE_URL_MANAGE_USERS: "https://manage-users-api-dev.hmpps.service.justice.gov.uk"
     API_BASE_URL_PRISON_API: "https://prison-api-dev.prison.service.justice.gov.uk"
     API_BASE_URL_PRISONER_SEARCH: "https://prisoner-search-dev.prison.service.justice.gov.uk"
-    API_BASE_URL_WHEREABOUTS: "https://whereabouts-api-dev.service.justice.gov.uk"
     API_BASE_URL_NOMIS_MAPPING: "https://nomis-sync-prisoner-mapping-dev.hmpps.service.justice.gov.uk"
     BVLS_FRONTEND_URL: "https://book-a-video-link-dev.prison.service.justice.gov.uk"
     FEATURE_EVENTS_SNS_ENABLED: true
