+++ helm_deploy/values-preprod.yaml
@@ -15,7 +15,6 @@ generic-service:
     API_BASE_URL_MANAGE_USERS: "https://manage-users-api-preprod.hmpps.service.justice.gov.uk"
     API_BASE_URL_PRISON_API: "https://prison-api-preprod.prison.service.justice.gov.uk"
     API_BASE_URL_PRISONER_SEARCH: "https://prisoner-search-preprod.prison.service.justice.gov.uk"
-    API_BASE_URL_WHEREABOUTS: "https://whereabouts-api-preprod.service.justice.gov.uk"
     API_BASE_URL_NOMIS_MAPPING: "https://nomis-sync-prisoner-mapping-preprod.hmpps.service.justice.gov.uk"
     BVLS_FRONTEND_URL: "https://book-a-video-link-preprod.prison.service.justice.gov.uk"
     FEATURE_EVENTS_SNS_ENABLED: true
