+++ helm_deploy/values-prod.yaml
@@ -12,7 +12,6 @@ generic-service:
     API_BASE_URL_MANAGE_USERS: "https://manage-users-api.hmpps.service.justice.gov.uk"
     API_BASE_URL_PRISON_API: "https://prison-api.prison.service.justice.gov.uk"
     API_BASE_URL_PRISONER_SEARCH: "https://prisoner-search.prison.service.justice.gov.uk"
-    API_BASE_URL_WHEREABOUTS: "https://whereabouts-api.service.justice.gov.uk"
     API_BASE_URL_NOMIS_MAPPING: "https://nomis-sync-prisoner-mapping.hmpps.service.justice.gov.uk"
     BVLS_FRONTEND_URL: "https://book-a-video-link.prison.service.justice.gov.uk"
     FEATURE_EVENTS_SNS_ENABLED: true
