+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsbookavideolinkapi/config/WebClientConfiguration.kt
@@ -17,7 +17,6 @@ class WebClientConfiguration(
   @Value("\${api.base.url.manage-users}") private val manageUsersBaseUri: String,
   @Value("\${api.base.url.prison-api}") val prisonApiBaseUri: String,
   @Value("\${api.base.url.prisoner-search}") val prisonerSearchBaseUri: String,
-  @Value("\${api.base.url.whereabouts}") val whereaboutsBaseUri: String,
   @Value("\${api.base.url.nomis-mapping}") val nomisMappingBaseUri: String,
   @Value("\${api.health-timeout:2s}") val healthTimeout: Duration,
   @Value("\${api.timeout:60s}") val timeout: Duration,
@@ -67,11 +66,4 @@ class WebClientConfiguration(
   @Bean
   fun prisonerSearchApiWebClient(authorizedClientManager: OAuth2AuthorizedClientManager, builder: WebClient.Builder) =
     builder.authorisedWebClient(authorizedClientManager, "prisoner-search", prisonerSearchBaseUri, timeout)
-
-  @Bean
-  fun whereaboutsApiWebClient(authorizedClientManager: OAuth2AuthorizedClientManager, builder: WebClient.Builder) =
-    builder.authorisedWebClient(authorizedClientManager, "whereabouts-api", whereaboutsBaseUri, timeout)
-
-  @Bean
-  fun whereaboutsApiHealthWebClient(builder: WebClient.Builder) = builder.healthWebClient(whereaboutsBaseUri, healthTimeout)
 }
