+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsbookavideolinkapi/service/events/DomainEvents.kt
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
 import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.common.toOffsetString
-import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.events.handlers.MigrateVideoBookingEvent
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
@@ -142,10 +141,6 @@ enum class DomainEventType(val eventType: String, val description: String = "") 
   PRISONER_MERGED("prison-offender-events.prisoner.merged") {
     override fun toInboundEvent(mapper: ObjectMapper, message: String) =
       mapper.readValue<PrisonerMergedEvent>(message)
-  },
-  MIGRATE_VIDEO_BOOKING("whereabouts-api.videolink.migrate") {
-    override fun toInboundEvent(mapper: ObjectMapper, message: String) =
-      mapper.readValue<MigrateVideoBookingEvent>(message)
   },
   PRISONER_VIDEO_APPOINTMENT_CANCELLED("prison-offender-events.prisoner.video-appointment.cancelled") {
     override fun toInboundEvent(mapper: ObjectMapper, message: String) = mapper.readValue<PrisonerVideoAppointmentCancelledEvent>(message)
