--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsbookavideolinkapi/service/migration/MigrationException.kt
@@ -1,9 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.migration
-
-/**
- * A migration exception cannot be recovered from i.e. there is an unrecoverable reason as to why this has occurred.
- *
- * In the event of a migration exception it will require further independent investigation e.g. looking at logs etc.
- */
-@Deprecated(message = "Can be removed when migration is completed")
-class MigrationException(message: String) : RuntimeException(message)
