+++ src/main/resources/application.yml
@@ -66,12 +66,6 @@ spring:
             client-secret: ${system.client.secret}
             authorization-grant-type: client_credentials
             scope: read
-          whereabouts-api:
-            provider: hmpps-auth
-            client-id: ${system.client.id}
-            client-secret: ${system.client.secret}
-            authorization-grant-type: client_credentials
-            scope: read
           nomis-mapping-api:
             provider: hmpps-auth
             client-id: ${system.client.id}
